# hr_website/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HR Website API",
    "version": "1.0.0",
    "description": "API documentation for the HR Website backend."
  },
  "components": {
    "schemas": {
...
      "Facility": {
        "type": "object",
        "required": [
          "id",
          "branch_id",
          "name",
          "code",
          "category",
          "quantity",
          "condition",
          "last_checked_at",
          "note",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "branch_id": {
            "type": "string",
            "format": "uuid",
            "description": "Chi nhánh sở hữu thiết bị (public.branches.id)."
          },
          "name": {
            "type": "string",
            "example": "Máy pha cà phê Breville"
          },
          "code": {
            "type": "string",
            "nullable": true,
            "description": "Mã tài sản / số serial trên tem thiết bị."
          },
          "category": {
            "type": "string",
            "enum": [
              "KITCHEN",
              "COLD_STORAGE",
              "FURNITURE",
              "ELECTRICAL",
              "CLEANING",
              "OTHER"
            ],
            "description": "Nhóm thiết bị (Khu pha chế, Bảo quản lạnh, ...)."
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "description": "Số lượng thiết bị cùng loại tại chi nhánh."
          },
          "condition": {
            "type": "string",
            "enum": [
              "GOOD",
              "FAIR",
              "MAINTENANCE",
              "BROKEN"
            ],
            "description": "Tình trạng: tốt, cần theo dõi, đang sửa chữa, hỏng (SCRUM-46 xử lý bảo trì)."
          },
          "last_checked_at": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Ngày kiểm tra gần nhất (YYYY-MM-DD)."
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FacilityWriteRequest": {
        "type": "object",
        "required": [
          "branch_id",
          "name"
        ],
        "properties": {
          "branch_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "code": {
            "type": "string",
            "maxLength": 60,
            "nullable": true
          },
          "category": {
            "type": "string",
            "enum": [
              "KITCHEN",
              "COLD_STORAGE",
              "FURNITURE",
              "ELECTRICAL",
              "CLEANING",
              "OTHER"
            ],
            "default": "OTHER"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9999,
            "default": 1
          },
          "condition": {
            "type": "string",
            "enum": [
              "GOOD",
              "FAIR",
              "MAINTENANCE",
              "BROKEN"
            ],
            "default": "GOOD"
          },
          "last_checked_at": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          }
        }
      },
      "ReverseGeocodeResult": {
        "type": "object",
        "required": [
          "address"
        ],
        "properties": {
          "address": {
            "type": "string",
            "nullable": true,
            "maxLength": 300,
            "description": "Địa chỉ đầy đủ của toạ độ (Nominatim/OpenStreetMap, tiếng Việt); null khi nhà cung cấp không có địa chỉ cho điểm đó.",
            "example": "128 Lê Lợi, Phường Bến Thành, Quận 1, TP. Hồ Chí Minh"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Mô tả lỗi.",
            "example": "Trường `question` là bắt buộc và phải là chuỗi không rỗng."
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/geo/reverse": {
      "get": {
        "summary": "Turn a coordinate into an address",
        "description": "Reverse geocoding for the branch form: the pin on the map becomes the \"Địa chỉ\" text, so the\noperator does not have to type it. Addresses come from Nominatim (OpenStreetMap) — the same\nproject the branch map draws its tiles from — and need no API key. The coordinate must be\ninside Vietnamese territory, exactly the rule the geofence itself enforces. `address` is\nnull when the provider knows no address for that point; `503` when it cannot be reached.\n",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "latitude",
            "required": true,
            "schema": {
              "type": "number",
              "minimum": 8.5436,
              "maximum": 23.3883
            },
            "description": "Vĩ độ của điểm cần tra cứu."
          },
          {
            "in": "query",
            "name": "longitude",
            "required": true,
            "schema": {
              "type": "number",
              "minimum": 102.0967,
              "maximum": 109.4944
            },
            "description": "Kinh độ của điểm cần tra cứu."
          }
        ],
        "responses": {
          "200": {
            "description": "The address of the coordinate, or null when the provider has none.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReverseGeocodeResult"
                }
              }
            }
          },
          "400": {
            "description": "The coordinates are missing, not numbers, out of range, or outside Vietnam.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The address lookup service is not available; type the address instead.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/facilities": {
      "get": {
        "summary": "List facilities",
        "description": "Returns the per-branch facilities visible to the authenticated caller under the active RLS policy, sorted by name. SCRUM-45.",
        "tags": [
          "Facilities"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "branch_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Restrict the list to one branch."
          }
        ],
        "responses": {
          "200": {
            "description": "The facility list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Facility"
                  }
                }
              }
            }
          },
          "400": {
            "description": "branch_id is not a UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The caller's role may not read the inventory (SCRUM-59, manager and above).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a facility",
        "description": "Registers one facility of a branch. Restricted to managers (`facility:manage`, SCRUM-59) and enforced by RLS.",
        "tags": [
          "Facilities"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacilityWriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The facility was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Facility"
                }
              }
            }
          },
          "400": {
            "description": "The request body is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The caller's role may not manage facilities (SCRUM-59), or they do not head that branch (SCRUM-61)."
          },
          "404": {
            "description": "The branch does not exist or is not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}